# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Gin + Vue3 的后台管理系统 API。\u003cbr/\u003e统一响应格式：\u003ccode\u003e{\"code\":0,\"msg\":\"ok\",\"data\":...}\u003c/code\u003e，code=0 成功，code=7 业务失败。\u003cbr/\u003e所有 \u003ccode\u003e/api/v1/system/*\u003c/code\u003e 接口需要 \u003ccode\u003eAuthorization: Bearer \u0026lt;token\u0026gt;\u003c/code\u003e。",
        "title": "go-admin API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8989",
    "basePath": "/",
    "paths": {
...
        "/api/v1/system/user": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.UserUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "新增用户",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.UserCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysUser"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/system/user/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.UserListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/system/user/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_lihongsheng_go-admin_server_dto_system.UserCreateReq": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_dto_system.UserListResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysUser"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_dto_system.UserUpdateReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role_ids": {
                    "description": "RoleIDs 为 nil 表示不调整角色；空切片表示清空角色",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_lihongsheng_go-admin_server_model_system.SysMenu": {
            "type": "object",
            "properties": {
                "api_rules": {
                    "description": "JSON 数组：[{\"path\":\"/api/...\",\"method\":\"GET\"},...]",
                    "type": "string"
                },
                "children": {
                    "description": "虚拟字段，仅用于 JSON 序列化时组装树",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysMenu"
                    }
                },
                "component": {
                    "description": "前端组件路径；catalog 用 \"Layout\"",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "hidden": {
                    "description": "是否隐藏",
                    "type": "boolean"
                },
                "icon": {
                    "description": "图标名（element-ui icon）",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "keep_alive": {
                    "description": "是否缓存页面",
                    "type": "boolean"
                },
                "name": {
                    "description": "路由 name；button 为按钮名",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父节点 ID，0 表示根",
                    "type": "integer"
                },
                "path": {
                    "description": "路由路径；catalog 如 /system，menu 如 user",
                    "type": "string"
                },
                "permission": {
                    "description": "权限码；button 必有值，如 user:add",
                    "type": "string"
                },
                "redirect": {
                    "description": "重定向路径",
                    "type": "string"
                },
                "sort": {
                    "description": "排序序号",
                    "type": "integer"
                },
                "title": {
                    "description": "显示标题",
                    "type": "string"
                },
                "type": {
                    "description": "catalog / menu / button",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_model_system.SysRole": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "default_router": {
                    "description": "登录后默认首页路由",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysMenu"
                    }
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_model_system.SysUser": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysRole"
                    }
                },
                "status": {
                    "description": "1启用 0禁用",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_utils_response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}